% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopdf.R
\name{geopdf}
\alias{geopdf}
\title{Convert raster file to geopdf}
\usage{
geopdf(src, outfile, verbose = TRUE)
}
\arguments{
\item{src}{input raster file}

\item{outfile}{output pdf file}

\item{verbose}{should the function be chatty? (default \code{TRUE})}
}
\value{
the path name of the output file
}
\description{
Convert raster file to geopdf
}
\examples{
file <- system.file("DEM.tif", package = "bluedotmapper")
geopdf(file, "file2.pdf")

}
